requirements.txt
loguru
toml
setuptools

opencv-python
kafka-python
rich